Switch doc build to github actions
