+++ .github/workflows/doc.yml
@@ -14,10 +14,6 @@ jobs:
         sudo apt install doxygen
         cmake -E make_directory ${{runner.workspace}}/build
 
-    - name: Configure
-      working-directory: ${{runner.workspace}}/build
-      run: cmake -DCMAKE_BUILD_TYPE=${{matrix.build_type}} $GITHUB_WORKSPACE
-
     - name: Build
       working-directory: ${{runner.workspace}}/build
-      run: cmake --build . --config ${{matrix.build_type}}
+      run: $GITHUB_WORKSPACE/support/build-docs.py
